### 2021-01-24 #2985 @whedon
[PRE REVIEW]: Nanoq: fast quality control for nanopore reads
**Submitting author:** @esteinig (<a href="http://orcid.org/0000-0001-5399-5353">Eike J. Steinig</a>)
**Repository:** <a href="https://github.com/esteinig/nanoq" target ="_blank">https://github.com/esteinig/nanoq</a>
**Version:** v0.2.1
