### 2021-01-24 #2985 @danielskatz
@esteinig - given the relatively small size of this submission, the editors will now discuss if it meet the JOSS [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review. You should hear back about the next steps in 1-2 weeks.
