### 2021-01-26 #2985 @luizirber
Thanks for the reviewers suggestions, @esteinig!

I'll start by pinging @bovee (as author of needletail) and @natir (since I know he has experimented with nanopore data before). Are you two willing to review `nanoq`?
