### 2021-01-26 #2985 @esteinig
Thanks for that, please let me know if there is anything that can be improved / optimised. It was our first foray into Rust, and the processing speed of the libaries is exceptional, even without much optimisation.

NB: Mean error function for read scores is costly at each iteration, but couldn't think of any faster way to implement it.
