### 2021-01-28 esteinig/nanoq#17 @natir
[JOSS Review] Paper comment
This relates to my JOSS review openjournals/joss-reviews#2991

First thanks for your work on this tools and publications

Some comments about paper:

- [x] The interest of have rust-bio and needltail parser in same time isn't clear in publication
- [x] Could you add memory usage in benchmark ?
- [x] I think it's easier to read result if stats and filter are present in separate Table
- [x] Which parser is use by nanoq in benchmark ?
- [x] Could you compare `rust-bio-stats sequence-stats` to nanoq stats functionality is very similar
- [x] Compare error rate, read quality, read length, across each filter method, could help to see which method is the best or if we can't see change between methods
