### 2021-01-28 esteinig/nanoq#18 @natir
[JOSS Review] Code quality comment
This relates to my JOSS review openjournals/joss-reviews#2991

First thanks for your work on this tools and publications

Some comments about code quality:

- [x] Add Continuous Integration could help to improve and keep an high quality of code. You can take inspiration from [rust-bio github action](https://github.com/rust-bio/rust-bio/blob/master/.github/workflows/rust.yml) don't hesitate to include test coverage
- [x] I notice some classic rust style error in code use [clippy](https://github.com/rust-lang/rust-clippy) could help to detect and fix them
- [x] To facilitate contribution could you split `main.rs` in some submodule and add documentation
