### 2021-01-28 esteinig/nanoq#19 @natir
[JOSS Review] Other comments
This relates to my JOSS review openjournals/joss-reviews#2991

First thanks for your work on this tools and publications

I put here the comments that I could not group with others:

- [x] Link to nanopore documentation about quality score require login to nanopore website, can you provide another link ?
- [x] Did you plan to replace your own conda by a bioconda recipe ? (I could do it for you)
