### 2021-02-01 esteinig/nanoq#19 @esteinig
@natir thanks very much for your comments and improvements - in the other threads as well. I will work on this tonight :)
