### 2021-02-02 esteinig/nanoq#19 @esteinig
1. Yes, good point! There was a discussion on this during development of `Nanostat` which might be relevant and not behind a community wall. Please let me know if this is appropriate: https://gigabaseorgigabyte.wordpress.com/2017/06/26/averaging-basecall-quality-scores-the-right-way/

2. Yes, absolutely. I will give it a shot but have had Michael's excellent help last time. I might have to come back to your generous offer
