### 2021-02-02 esteinig/nanoq#17 @esteinig
Quick response before working on changes on `joss` branch:

1. Good point, I ended up using the Needletail parser as it currently supports `fasta/q` out of the box. I saw you are working on this for `rust-bio` for an upcoming release? Do you think it is worth including both?
2. Yes, absolutely.
3. Easily done, thanks.
4. It's the default Needletail parser, I will make that more clear
5. Ohhh very cool, I was not aware of this! Yep can be done, do you want to see it in the benchmarks as well?
6. Yep, absolutely - you may have seen the other thread on `Filtlong` differences as well

Thanks very much!
