### 2021-02-02 esteinig/nanoq#18 @esteinig
Again, some comments before working on the `joss` branch:

1. Yep can implement some CI (I think)
2. Oh cool, thanks for the tip, will lint the code
3. Yep will do!
