### 2021-02-02 esteinig/nanoq#17 @natir
Quick response to some quick response:

1 -> We agree that all you need is a fast·a·q parser ? Needletail support fast·a·q out of box, with gzip, bzip, lzma compression or not, support multi-thread and it's faster than rust-bio. Why did you want keep rust-bio parser ?
5 -> Yes sure !
