### 2021-02-02 esteinig/nanoq#19 @natir
1. it's ok
2. If you want you can take inspiration from my package [yacrd](https://github.com/bioconda/bioconda-recipes/tree/master/recipes/yacrd) or [fpa](https://github.com/bioconda/bioconda-recipes/tree/master/recipes/fpa) don't hesitate to tag me in your bioconda pull request, I would review it with pleasure.
