### 2021-02-02 esteinig/nanoq#17 @esteinig
That was quick!

Mainly because I really like rust-bio, but I see your point :) I should contribute the filter function to rust-bio-tools instead
