### 2021-02-02 esteinig/nanoq#17 @natir
I like rust-bio too but for your needs isn't the crate you need
