### 2021-02-02 esteinig/nanoq#17 @esteinig
Indeed, `needletail` is pretty amazing
