### 2021-02-11 #2991 @luizirber
We are in the middle of a pandemic, so whedon's comment might sound a bit demanding, but the intentions are good, just so we know how the review is going =]

@esteinig do you mind [opening a PR](https://github.com/esteinig/nanoq/compare/joss?expand=1) with the `joss` branch, so we have better visibility on the changes you're making to address Roderick and Pierre's comments? You don't need to merge the PR yet, but it is nicer to see changes than having to find the branch in the nanoq repo.
