### 2021-03-25 #2991 @arfon
> Just checking with @openjournals/joss-eics if there is any specific whedon command or tag to be added to show that this submission is paused, or just leave as-is.

We just add the `paused` label in situations like this. I'll add it now...
