### 2021-04-10 88d871d
Update info regarding conda installation
nanoq is now available at Bioconda so I updated the documentation to reflect that.
