### 2021-09-23 9826256
stable re-write [#18]; fast mode [#21]; bug fix [#21]; rm complexity [#17]
