+++ src/utils.rs
@@ -1,0 +1,447 @@
+use indoc::eprintdoc;
+use thiserror::Error;
+use anyhow::Result;
+
+/// A collection of custom errors relating to the utility components for this package.
+#[derive(Error, Debug, PartialEq)]
+pub enum UtilityError {
+
+    /// Indicates an invalid verbosity for summary output
+    #[error("{0} is not a valid level of verbosity")]
+    InvalidVerbosity(String),
+
+}
+
+/// Create a read set object
+///
+/// Read set objects are mutable to allow
+/// sorting of read length and quality vectors
+#[derive(Debug)]
+pub struct ReadSet {
+    pub read_lengths: Vec<u32>,
+    pub read_qualities: Vec<f32>,
+}
+
+impl ReadSet {
+    /// Print a summary of the read set to stderr
+    ///
+    /// * `detail` - verbosity of summary message
+    ///     * 0: standard output without headers
+    ///     * 1: standard output with pretty headers
+    ///     * 2: add length and quality thresholds
+    ///     * 3: add top ranked read statistics
+    ///
+    /// * `top` - show top ranking lengths and qualities
+    ///
+    /// # Example
+    /// 
+    /// ```rust
+    /// let mut read_set = ReadSet {
+    ///     read_lengths: !vec[10, 100, 1000],
+    ///     read_qualities: !vec[10.0, 11.0, 12.0],
+    /// }
+    /// read_set.summary(0, 3)
+    /// read_set.summary(1, 3)
+    /// read_set.summary(2, 3)
+    /// read_set.summary(3, 3)
+    /// ```
+    pub fn summary(&mut self, verbosity: &u8, top: usize) -> Result<(), UtilityError> {
+
+        let (min_length, max_length) = self.range_length();
+
+        match verbosity {
+            0 => {
+                eprintdoc! {
+                    "{reads} {bases} {n50} {longest} {shortest} {mean} {median} {meanq:.1} {medianq:.1}\n",
+                    reads = self.reads(),
+                    bases = self.bases(),
+                    n50 = self.n50(),
+                    longest = max_length,
+                    shortest = min_length,
+                    mean = self.mean_length(),
+                    median = self.median_length(),
+                    meanq = self.mean_quality(),
+                    medianq = self.median_quality(),
+                }
+                Ok(())
+            },
+            1 | 2 | 3 => {
+                eprintdoc! {"\n
+                    Nanoq Read Summary
+                    ====================
+                    
+                    Number of reads:      {reads}
+                    Number of bases:      {bases}
+                    N50 read length:      {n50}
+                    Longest read:         {longest} 
+                    Shortest read:        {shortest}
+                    Mean read length:     {mean}
+                    Median read length:   {median} 
+                    Mean read quality:    {meanq:.2} 
+                    Median read quality:  {medianq:.2}
+                    ",
+                    reads = self.reads(),
+                    bases = self.bases(),
+                    n50 = self.n50(),
+                    longest = max_length,
+                    shortest = min_length,
+                    mean = self.mean_length(),
+                    median = self.median_length(),
+                    meanq = self.mean_quality(),
+                    medianq = self.median_quality(),
+                }
+                if *verbosity > 1  {
+                    self.print_thresholds();
+                } 
+                if *verbosity > 2 {
[…]
