### 2021-09-23 6bf2049
add blog ref [#19]; bump version; restruct readme
