+++ README.md
@@ -1,14 +1,14 @@
 # nanoq <a href='https://github.com/esteinig'><img src='docs/nanoq.png' align="right" height="270" /></a>
 
 ![](https://img.shields.io/badge/lang-rust-black.svg)
-![](https://img.shields.io/badge/version-0.2.1-purple.svg)
+![](https://img.shields.io/badge/version-0.7.0-purple.svg)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3707754.svg)](https://doi.org/10.5281/zenodo.3707754)
 
 Minimal but speedy quality control for nanopore reads.
 
 ## Overview
 
-**`v0.2.1`**
+**`v0.7.0`**
 
 - [Purpose](#purpose)
 - [Install](#install)
@@ -23,9 +23,7 @@ Minimal but speedy quality control for nanopore reads.
 
 ## Purpose
 
-Basic sequence quality control and computation of summary statistics can be a bit slow due to bottlenecks in read parsing. `Nanoq` attempts to perform these operations on `fastx` files using the `needletail` and `rust-bio` libraries with either a single-pass operation for defaulty summary statistics and filtering, or a two-pass operation enabling filtering methods similar to [`Filtlong`](https://github.com/rrwick/Filtlong).
-
-Quality scores are computed for basecalls from nanopore sequencing data, as outlined in the [technical documentation](https://community.nanoporetech.com/technical_documents/data-analysis/).
+`Nanoq` implements fast sequence read and quality filtering and produces simpel reports. Quality scores are computed for basecalls from nanopore sequencing data, as outlined in the [technical documentation](https://community.nanoporetech.com/technical_documents/data-analysis/) or in this [blog post](https://gigabaseorgigabyte.wordpress.com/2017/06/26/averaging-basecall-quality-scores-the-right-way/).
 
 ## Install
 
@@ -42,99 +40,100 @@ cargo install nanoq
 Currently on this channel but will be in `BioConda`:
 
 ```
-conda install -c conda-forge -c esteinig nanoq=0.2.1
+conda install -c conda-forge -c esteinig nanoq=0.7.0
 ```
 
 #### `Docker`
 
-`Docker` image is based on the light `Alpine OS` (~ 21 MB)
+`Docker` image is based on `Alpine OS` (~ 20 MB)
 
 ```
 docker pull esteinig/nanoq:latest
 ```
 
-#### `Singularity`
-
-```
-singularity pull docker://esteinig/nanoq:latest
-```
-
 ## Usage
 
-Tests can be run within the `nanoq` repository:
-
-```
-cargo test
-```
-
-`Nanoq` accepts a file or stream of sequence reads in `fast{a/q}` and compressed formats on `stdin`:
+`Nanoq` accepts a file (`--input`) or stream (`stdin`) of reads in `fast{a,q}.{gz,bz2,xz}` format and outputs reads to file (`--output`) or stream (`stdout`).
 
 ```bash
-cat test.fq | nanoq
+nanoq -i test1.fq.gz -o test2.fq
+cat test1.fq.gz | nanoq > test2.fq
 ```
 
-Reads filtered by minimum read length (`--length`) and mean read quality (`--quality`) are output to `stdout`:
+Output compression is inferred from file extensions (`gz`, `bz2`, `lzma`).
 
 ```bash
-cat test.fq | nanoq -l 1000 -q 10 > reads.fq 
+nanoq -i test1.fq -o test2.fq.gz
 ```
 
[…]
