+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nanoq"
-version = "0.2.1"
+version = "0.7.0"
 authors = ["esteinig <eike.steinig@unimelb.edu.au>", "ljmcoin <lachlan.coin@unimelb.edu.au>"]
 description = "Minimal but speedy quality control and summaries of nanopore reads"
 documentation = "https://github.com/esteinig/nanoq"
@@ -16,6 +16,9 @@ include = [
     "Cargo.toml"
 ]
 
+[badges]
+maintenance = { status = "actively-developed" }
+
 [dependencies]
 structopt = "0.3"
 clap = "2.33.0"
@@ -24,3 +27,8 @@ indoc = "1.0"
 needletail = "0.4.1"
 thiserror = "1.0"
 niffler = "2.3"
+
+[dev-dependencies]
+assert_cmd = "0.10"
+predicates = "1"
+tempfile = "3.1.0"
