+++ README.md
@@ -23,7 +23,7 @@ Minimal but speedy quality control for nanopore reads.
 
 ## Purpose
 
-`Nanoq` implements fast sequence read and quality filtering and produces simpel reports. Quality scores are computed for basecalls from nanopore sequencing data, as outlined in the [technical documentation](https://community.nanoporetech.com/technical_documents/data-analysis/) or in this [blog post](https://gigabaseorgigabyte.wordpress.com/2017/06/26/averaging-basecall-quality-scores-the-right-way/).
+`Nanoq` implements fast sequence read and quality filtering and produces simple summary reports. Quality scores are computed for basecalls from nanopore sequencing data, as outlined in the [technical documentation](https://community.nanoporetech.com/technical_documents/data-analysis/) or in this [blog post](https://gigabaseorgigabyte.wordpress.com/2017/06/26/averaging-basecall-quality-scores-the-right-way/).
 
 ## Install
 
