### 2021-09-24 0011b40
add code docs [#18]; redo benchmarks [#17]; rewrite paper [#17]
