+++ paper/paper.md
@@ -25,17 +25,17 @@ bibliography: paper.bib
 
 # Summary
 
-Nanopore sequencing is now routinely integrated in a variety of genomics applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. As a consequence, the amount of nanopore sequence data in the public domain has increased rapidly in the last few years. One of the first steps in any workflow is to assess the quality of reads and obtain basic summary statistics after basecalling raw nanopore signal, as well as to filter fragmented or low quality reads. [`NanoPack`](https://github.com/wdecoster/nanopack) (`biopython` parser) [@nanopack], [`Filtlong`](https://github.com/rrwick/Filtlong) ([`Klib`](https://github.com/attractivechaos/klib) parser) and [`MinIONQC`](https://github.com/roblanf/minion_qc/blob/master/README.md) (summary file parser) [@minionqc] are common tools used to filter and obtain summary statistics from nanopore reads. However, these tools may be bottlenecked with regards to read parsing (`NanoPack`, `Filtlong`), not immediately applicable due to reliance on summary files (`MinIONQC`) or focus on more complex filters, data exploration and visualization. We therefore wrote `nanoq`, a minimal command line application to accelerate quality control and read summaries for large nanopore data sets.
+Nanopore sequencing is now routinely used in a variety of applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. As a consequence, the amount of nanopore sequence data in the public domain has increased rapidly in the last few years. One of the first steps in any workflow is to assess the quality of reads and obtain basic summary statistics after basecalling raw nanopore signal data, as well as to filter fragmented or low quality reads. [`NanoPack`](https://github.com/wdecoster/nanopack) (`biopython` parser) [@nanopack], [`Filtlong`](https://github.com/rrwick/Filtlong) ([`Klib`](https://github.com/attractivechaos/klib) parser) and [`MinIONQC`](https://github.com/roblanf/minion_qc/blob/master/README.md) (summary file parser) [@minionqc] are common tools used to filter and obtain summary statistics from nanopore reads. However, these tools may be bottlenecked during read iteration (`NanoPack`, `Filtlong`), not immediately applicable due to reliance on summary files (`MinIONQC`) or implement complex filters and data visualization. We therefore wrote `nanoq`, a minimal command line application to accelerate quality control and read summaries for large nanopore data sets.
 
 # Statement of need
 
-A common practice for quality control and filtering of nanopore reads is to use a sequencing summary index to speed up iteration over millions of reads, which requires access to these files. With increasing throughput on scalable nanopore platforms like `GridION` or `PromethION`, fast quality control of sequence reads and the ability to generate summary statistics on-the-fly are required. `Nanoq` is highly competitive in processing speed (see benchmarks) and can be effectively applied to nanopore data from the public domain, as part of automated pipelines, in streaming applications, or from the command line to check on the progress of active sequencing runs.
[…]
