+++ src/needlecast.rs
@@ -1,43 +1,60 @@
 
 
+
+use needletail::{FastxReader, parse_fastx_file, parse_fastx_stdin};
+use std::io::{BufWriter, Write};
+use std::io::{stdout, sink};
+use thiserror::Error;
+use std::fs::File;
+
 use crate::cli::Cli;
 use crate::cli::CompressionExt;
-use std::fs::File;
-use std::io::{BufWriter, Write};
-use needletail::{FastxReader, parse_fastx_file, parse_fastx_stdin};
-use thiserror::Error;
-use std::io::{stdout, sink};
+
 
 // Niffler output compression adopted from Michael B. Hall - Rasusa (https://github.com/mbhall88/rasusa)
-
 
 /// A collection of custom errors relating to the working with fastx files.
 #[derive(Error, Debug)]
 pub enum FastxError {
 
     /// Indicates that a sequence record could not be parsed.
-    #[error("Failed to parse record")]
+    #[error("Failed to parse sequence record")]
     ParseError {
         source: needletail::errors::ParseError,
     },
 
 }
 
+/// NeedleCast object
+///
+/// Basically a minimal wrapper around Needletail
+/// that implements parsing and filtering based on
+/// read length and quality
 pub struct NeedleCast {
     reader: Box<dyn FastxReader>,
     writer: Box<dyn Write>,
-
 }
 
 impl NeedleCast {
-
+    /// Create a new NeedleCast instance
+    ///
+    /// Given the command line interface object,
+    /// create a new instance that parses its
+    /// arguments and instantiates the
+    /// reader and writer
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// let args = Cli::from_iter(&["nanoq"]);
+    /// let nc = NeedleCast::new(&args);
+    /// ```
     pub fn new(args: &Cli) -> Self {
 
         let reader = match &args.input {
             Some(file) => parse_fastx_file(file).expect("failed to get file reader"),
             None => parse_fastx_stdin().expect("failed to get stdin reader"),
         };
-
         let writer = match &args.output {
             None => {
                 if args.stats {
@@ -60,26 +77,40 @@ impl NeedleCast {
                     None => niffler::Format::from_path(&output),
                     Some(f) => f,
                 };
-
                 niffler::get_writer(
                     file_handle, fmt, args.compress_level
                 ).expect("failed to get compressed file writer")
             }
         };
-
         NeedleCast {
             reader, 
             writer
         }
-    
     }
+    /// Filter reads and store lengths and qualities
+    ///
+    /// Given filtering parameters, iterate over reads
+    /// and compute average read quality (fastq) records.
+    /// Read lengths and qualities are stored in vectors
+    /// and returned if no errors are raised
+    /// 
+    /// # Errors
+    /// 
+    /// If the sequence record cannot be parsed a variant
+    /// of the `Needletail` `FastxError` is returned 
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// let args = Cli::from_iter(&["nanoq"]);
+    /// let nc = NeedleCast::new(&args);
+    /// nc.filter(0, 0, 0);
+    /// ```
     pub fn filter(&mut self, min_length: u32, max_length: u32, min_quality: f32) -> Result<(Vec<u32>, Vec<f32>), FastxError> {
         let mut read_lengths: Vec<u32> = vec![];
         let mut read_qualities: Vec<f32> = vec![];
 
-        
         let max_length: u32 = if max_length <= 0 { u32::MAX } else { max_length };
-
         while let Some(record) = self.reader.next() {
             match record {
                 Ok(rec) => {
[…]
