### 2021-09-24 esteinig/nanoq#17 @esteinig
@natir my apologies for the long pause, and thanks for your patience. Some uhm interesting times.

I have rewritten the code base and will go through your review after adding tests over the weekend. Thanks again for the suggestions, they helped improve `nanoq` a lot.
