### 2021-09-26 58d0f9b
add unit tests [#18]; add rbt benchmark [#17]; redo benchmarks [#17]
