+++ paper/paper.md
@@ -25,17 +25,19 @@ bibliography: paper.bib
 
 # Summary
 
-Nanopore sequencing is now routinely used in a variety of applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. As a consequence, the amount of nanopore sequence data in the public domain has increased rapidly in the last few years. One of the first steps in any workflow is to assess the quality of reads and obtain basic summary statistics after basecalling raw nanopore signal data, as well as to filter fragmented or low quality reads. [`NanoPack`](https://github.com/wdecoster/nanopack) (`biopython` parser) [@nanopack], [`Filtlong`](https://github.com/rrwick/Filtlong) ([`Klib`](https://github.com/attractivechaos/klib) parser) and [`MinIONQC`](https://github.com/roblanf/minion_qc/blob/master/README.md) (summary file parser) [@minionqc] are common tools used to filter and obtain summary statistics from nanopore reads. However, these tools may be bottlenecked during read iteration (`NanoPack`, `Filtlong`), not immediately applicable due to reliance on summary files (`MinIONQC`) or implement complex filters and data visualization. We therefore wrote `nanoq`, a minimal command line application to accelerate quality control and read summaries for large nanopore data sets.
+Nanopore sequencing is now routinely used in a variety of genomics applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. One of the first steps in many workflows is to assess the quality of reads and obtain basic summary statistics after basecalling signal data, as well as to filter fragmented or low quality reads. With increasing throughput on scalable nanopore platforms like `GridION` or `PromethION`, fast quality control of sequence reads and the ability to generate summary statistics on-the-fly are required. `Nanoq` is 2-3x faster than `rust-bio-tools` and up to 400x faster than commonly used alternatives (see benchmarks) and offers nanopore-specific read quality scores, common filtering options and output compression. `Nanoq` can therefore be effectively applied to nanopore data from the public domain, as part of automated pipelines, in streaming applications, or to check on the progress of active sequencing runs.
 
 # Statement of need
 
-A common practice for quality control and filtering of nanopore reads is to use a sequencing summary index to speed up iteration over millions of reads, which requires access to the original files. With increasing throughput on scalable nanopore platforms like `GridION` or `PromethION`, fast quality control of sequence reads and the ability to generate summary statistics on-the-fly are required. `Nanoq` is highly competitive in processing speed (see benchmarks) and can be effectively applied to nanopore data from the public domain, as part of automated pipelines, in streaming applications, or from the command line to check on the progress of active sequencing runs.
[…]
