+++ src/utils.rs
@@ -1,6 +1,8 @@
 use anyhow::Result;
 use indoc::eprintdoc;
 use thiserror::Error;
+use std::ffi::OsStr;
+use std::path::Path;
 
 /// A collection of custom errors relating to the utility components for this package.
 #[derive(Error, Debug, PartialEq)]
@@ -12,6 +14,27 @@ pub enum UtilityError {
 
 }
 
+// Adopted from Michael B. Hall - Rasusa (https://github.com/mbhall88/rasusa)
+
+pub trait CompressionExt {
+    fn from_path<S: AsRef<OsStr> + ?Sized>(p: &S) -> Self;
+}
+
+/// Attempts to infer the compression type from the file extension. 
+/// If the extension is not known, then Uncompressed is returned.
+impl CompressionExt for niffler::compression::Format {
+    fn from_path<S: AsRef<OsStr> + ?Sized>(p: &S) -> Self {
+        let path = Path::new(p);
+        match path.extension().map(|s| s.to_str()) {
+            Some(Some("gz")) => Self::Gzip,
+            Some(Some("bz") | Some("bz2")) => Self::Bzip,
+            Some(Some("lzma")) => Self::Lzma,
+            _ => Self::No,
+        }
+    }
+}
+
+
 /// ReadSet object
 ///
 /// Read set objects are mutable to allow
@@ -22,11 +45,29 @@ pub enum UtilityError {
 /// 
 #[derive(Debug)]
 pub struct ReadSet {
-    pub read_lengths: Vec<u32>,
-    pub read_qualities: Vec<f32>,
+    read_lengths: Vec<u32>,
+    read_qualities: Vec<f32>,
 }
 
 impl ReadSet {
+    /// Create a new ReadSet instance
+    ///
+    /// Given the verctors of read lengths and 
+    /// qualities return a mutable ReadSet
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// let mut read_set = ReadSet::new(
+    ///     vec![10, 100, 1000], vec![10.0, 11.0, 12.0]
+    /// )
+    /// ```
+    pub fn new(read_lengths: Vec<u32>, read_qualities: Vec<f32>) -> Self {
+        ReadSet {
+            read_lengths,
+            read_qualities,
+        }
+    }
     /// Print a summary of the read set to stderr
     ///
     /// * `verbosity` - detail of summary message
@@ -41,25 +82,24 @@ impl ReadSet {
     /// # Example
     /// 
     /// ```rust
-    /// let mut read_set = ReadSet {
-    ///     read_lengths: !vec[10, 100, 1000],
-    ///     read_qualities: !vec[10.0, 11.0, 12.0],
-    /// }
+    /// let mut read_set = ReadSet::new(
+    ///     !vec[10, 100, 1000], !vec[10.0, 11.0, 12.0]
+    /// )
     /// read_set.summary(0, 3);
     /// ```
-    pub fn summary(&mut self, verbosity: &u8, top: usize) -> Result<(), UtilityError> {
+    pub fn summary(&mut self, verbosity: &u64, top: usize) -> Result<(), UtilityError> {
 
-        let (min_length, max_length) = self.range_length();
+        let length_range = self.range_length();
 
         match verbosity {
-            0 => {
+            &0 => {
                 eprintdoc! {
                     "{reads} {bases} {n50} {longest} {shortest} {mean} {median} {meanq:.1} {medianq:.1}\n",
                     reads = self.reads(),
                     bases = self.bases(),
                     n50 = self.n50(),
-                    longest = max_length,
-                    shortest = min_length,
+                    longest = length_range[0],
+                    shortest = length_range[1],
                     mean = self.mean_length(),
                     median = self.median_length(),
                     meanq = self.mean_quality(),
[…]
