+++ src/cli.rs
@@ -1,7 +1,6 @@
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use structopt::StructOpt;
 use thiserror::Error;
-use std::ffi::OsStr;
 
 /// Read filters and summary reports for nanopore data
 #[derive(Debug, StructOpt)]
@@ -43,7 +42,7 @@ pub struct Cli {
 
     /// Minimum average read quality filter (Q).
     #[structopt(
-        short,
+        short="q",
         long,
         value_name = "FLOAT", 
         default_value = "0" 
@@ -54,9 +53,9 @@ pub struct Cli {
     #[structopt(
         short,
         long,
-        parse(from_occurrences)
+        parse(from_occurrences = parse_verbosity)
     )]
-    pub verbose: u8,
+    pub verbose: u64,
 
     /// Number of top reads in verbose summary.  
     #[structopt(
@@ -124,28 +123,18 @@ pub enum CliError {
 }
 
 
-// Adopted from Michael B. Hall - Rasusa (https://github.com/mbhall88/rasusa)
-
-pub trait CompressionExt {
-    fn from_path<S: AsRef<OsStr> + ?Sized>(p: &S) -> Self;
-}
-
-/// Attempts to infer the compression type from the file extension. 
-/// If the extension is not known, then Uncompressed is returned.
-impl CompressionExt for niffler::compression::Format {
-    fn from_path<S: AsRef<OsStr> + ?Sized>(p: &S) -> Self {
-        let path = Path::new(p);
-        match path.extension().map(|s| s.to_str()) {
-            Some(Some("gz")) => Self::Gzip,
-            Some(Some("bz") | Some("bz2")) => Self::Bzip,
-            Some(Some("lzma")) => Self::Lzma,
-            _ => Self::No,
-        }
+/// Utility function to parse verbosity occurences
+///
+/// Up to three verbosity flags are allowed (-vvv), if more
+/// are specified (-vvvv) the highest allowed value is returned
+pub fn parse_verbosity(v: u64) -> u64 {
+    match v {
+        0 | 1 | 2 | 3 => v,
+        _ => 3,
     }
 }
 
-
-/// Utility function to parse compression format raising error if no valid format is provided
+/// Utility function to parse compression format 
 fn parse_compression_format(s: &str) -> Result<niffler::compression::Format, CliError> {
     match s {
         "b" | "B" => Ok(niffler::Format::Bzip),
@@ -156,7 +145,7 @@ fn parse_compression_format(s: &str) -> Result<niffler::compression::Format, Cli
     }
 }
 
-/// Utility function to validate compression level is in allowed range
+/// Utility function to parse and validate compression level
 #[allow(clippy::redundant_clone)]
 fn parse_compression_level(s: &str) -> Result<niffler::Level, CliError> {
     let lvl = match s.parse::<u8>() {
@@ -174,21 +163,167 @@ fn parse_compression_level(s: &str) -> Result<niffler::Level, CliError> {
     Ok(lvl)
 }
 
+
 #[cfg(test)]
 mod tests {
     use super::*;
 
-
     #[test]
-    fn no_input_file_given_raises_error() {
-        let passed_args = vec!["nanoq", "-l", "1000"];
+    fn invalid_compression_format() {
+        let passed_args = vec!["nanoq", "-O", "t"];
         let args: Result<Cli, clap::Error> = Cli::from_iter_safe(passed_args);
 
         let actual = args.unwrap_err().kind;
-        let expected = clap::ErrorKind::MissingRequiredArgument;
+        let expected = clap::ErrorKind::InvalidValue;
 
         assert_eq!(actual, expected)
     }
 
+    #[test]
+    fn invalid_compression_level() {
[…]

[…]
