### 2021-09-27 2b430b9
add final benchmarks [#17]
