+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Nanoq: fast quality control for nanopore reads'
+title: 'Nanoq: ultra-fast quality control for nanopore reads'
 tags:
   - ont
   - nanopore
@@ -25,13 +25,11 @@ bibliography: paper.bib
 
 # Summary
 
-Nanopore sequencing is now routinely used in a variety of genomics applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. One of the first steps in many workflows is to assess the quality of reads and obtain basic summary statistics after basecalling signal data, as well as to filter fragmented or low quality reads. With increasing throughput on scalable nanopore platforms like `GridION` or `PromethION`, fast quality control of sequence reads and the ability to generate summary statistics on-the-fly are required. `Nanoq` is 2-3x faster than `rust-bio-tools` and up to 400x faster than commonly used alternatives (see benchmarks) and offers nanopore-specific read quality scores, common filtering options and output compression. `Nanoq` can therefore be effectively applied to nanopore data from the public domain, as part of automated pipelines, in streaming applications, or to check on the progress of active sequencing runs.
+Nanopore sequencing is now routinely used in a variety of genomics applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. One of the first steps in many workflows is to assess the quality of reads and obtain basic summary statistics after basecalling signal data, as well as to filter fragmented or low quality reads. With increasing throughput on scalable nanopore platforms like `GridION` or `PromethION`, fast quality control of sequence reads and the ability to generate summary statistics on-the-fly are required. When computing minimum summary statistics, `Nanoq` is ~2x faster than `rust-bio-tools` (~5 Gbp and >1 million reads per second) and up to ~440x faster than commonly used alternatives. It offers nanopore-specific quality scores, read filtering options, output compression and can be applied to data from the public domain, as part of automated pipelines, in streaming applications, or to quickly check active sequencing runs.
 
 # Statement of need
 
  [`NanoPack`](https://github.com/wdecoster/nanopack) (`biopython`) [@nanopack], [`Filtlong`](https://github.com/rrwick/Filtlong) ([`Klib`](https://github.com/attractivechaos/klib)) and [`MinIONQC`](https://github.com/roblanf/minion_qc/blob/master/README.md) (basecalling summary) [@minionqc] are common tools used to filter and obtain summary statistics from nanopore reads. However, these tools may be bottlenecked during read iteration (`NanoPack`, `Filtlong`), not immediately applicable due to reliance on basecalling summary files (`MinIONQC`) or implement complex filters and data visualization for research applications. We wrote `nanoq` to accelerate quality control and summary statistics for large nanopore data sets.
-
-
 
 # Applications
 
@@ -77,7 +75,6 @@ Read set summary statistics are output to `stderr`:
 * mean and median read length
[…]
