+++ paper/plot_benchmarks.py
@@ -8,62 +8,179 @@ from matplotlib import pyplot as plt
 def main():
 
     data = read_data()
-    plot_data(data)
+    summary = get_data_summary(data=data)
+
+    with pandas.option_context('display.max_rows', None):
+        print(data)
+        print(summary)
+
+    summary.to_csv('benchmark_summary.tsv', sep='\t', index=False)
+
+    plot_data(data=data)
+
+def get_data_summary(data: pandas.DataFrame) -> pandas.DataFrame:
+
+    data['mem'] = data['mem'] / 1000  # MB
+    data = data.sort_values(['task', 'file', 'tool'])
+
+    # Summary per category
+    summary = []
+    for task, _data in data.groupby("task"):
+        for file, __data in _data.groupby("file"):
+            for tool, ___data in __data.groupby("tool"):
+                mean_seconds = ___data['time'].mean()
+                std_seconds = ___data['time'].std()
+                mean_mem = ___data['mem'].mean()
+                std_mem = ___data['mem'].std()
+                summary.append([task, file, tool, mean_seconds, std_seconds, mean_mem, std_mem])
+
+    return pandas.DataFrame(summary, columns=['task', 'file', 'tool', 'time', 'time_sd', 'mem', 'mem_sd'])
+
 
 def read_data() -> pandas.DataFrame:
 
     data = []
-    for f in Path('replicate_benchmarks').glob("*"):
-        df = pandas.read_csv(f, header=None, sep="\t", names=['time'])
+    for f in Path('benchmarks').glob("*"):
+        df = pandas.read_csv(f, header=None, sep=" ", names=['time', 'mem'])
 
-        tool, ftype, mode = f.name.split("_")
-        df['tool'] = [tool for _ in df.iterrows()]
-        df['ftype'] = [ftype for _ in df.iterrows()]
-        df['mode'] = [mode for _ in df.iterrows()]
+        tool, task, mode = f.name.split("_")
 
-        data.append(df)
+        if tool == "nanostat8":
+            tool = "nanostat-t8"
+        elif tool == "nanoqf":
+            tool = "nanoq-fast"
+        elif tool == "rbt":
+            tool = "rust-bio-tools"
+        
+
+        df1 = df.iloc[1:11, :].copy()  # cold start
+        df1['tool'] = [tool for _ in df1.iterrows()]
+        df1['task'] = [task for _ in df1.iterrows()]
+        df1['file'] = [f'zymo_{mode}' for _ in df1.iterrows()]
+
+        data.append(df1)
+
+        if f.name.endswith('fq'):
+            df2 = df.iloc[12:, :].copy()
+            df2['tool'] = [tool for _ in df2.iterrows()]
+            df2['task'] = [task for _ in df2.iterrows()]
+            df2['file'] = ['zymo_full' for _ in df2.iterrows()]
+
+            data.append(df2)
+            
 
     return pandas.concat(data)
 
 def plot_data(data: pandas.DataFrame) -> None:
 
-    fig, axes = plt.subplots(
-        nrows=1, ncols=2, figsize=(
-            2 * 7, 1 * 4.5
+    # Zymo
+
+    fig1, axes1 = plt.subplots(
+        nrows=2, ncols=2, figsize=(
+            2 * 7, 2 * 4.5
         )
     )
 
[…]
