### 2021-09-27 e4d8709
add ci tools [#18]; add paper sections
