+++ README.md
@@ -4,7 +4,7 @@
 ![](https://img.shields.io/badge/version-0.7.0-purple.svg)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3707754.svg)](https://doi.org/10.5281/zenodo.3707754)
 
-Minimal but speedy quality control for nanopore reads.
+Minimal but ultra-speedy quality control for nanopore reads.
 
 ## Overview
 
@@ -29,23 +29,17 @@ Minimal but speedy quality control for nanopore reads.
 
 #### `Cargo`
 
-If you have [`Rust`](https://www.rust-lang.org/tools/install) and `Cargo` installed:
-
 ```
 cargo install nanoq
 ```
 
 #### `Conda`
 
-Currently on this channel but will be in `BioConda`:
-
 ```
-conda install -c conda-forge -c esteinig nanoq=0.7.0
+conda install -c bioconda nanoq=0.7.0
 ```
 
 #### `Docker`
-
-`Docker` image is based on `Alpine OS` (~ 20 MB)
 
 ```
 docker pull esteinig/nanoq:latest
@@ -53,7 +47,7 @@ docker pull esteinig/nanoq:latest
 
 ## Usage
 
-`Nanoq` accepts a file (`-i / --input`) or stream (`stdin`) of reads in `fast{a,q}.{gz,bz2,xz}` format and outputs reads to file (`-o / --output`) or stream (`stdout`).
+`Nanoq` accepts a file (`-i`) or stream (`stdin`) of reads in `fast{a,q}.{gz,bz2,xz}` format and outputs reads to file (`-o`) or stream (`stdout`).
 
 ```bash
 nanoq -i test1.fq.gz -o test2.fq
@@ -66,26 +60,26 @@ Output compression is inferred from file extensions (`gz`, `bz2`, `lzma`).
 nanoq -i test1.fq -o test2.fq.gz
 ```
 
-Output compression can be specified manually with `-O / --output-type` and `-c / --compress-level`.
+Output compression can be specified manually with `-O` and `-c`.
 
 ```bash
 nanoq -i test1.fq -O g -c 9 -o test2.fq.cmp
 ```
 
-Reads can be filtered by minimum read length (`-l / --min-len`), maximum read length (`-m / --max-len`) or average read quality (`-q / --min-qual`).
+Reads can be filtered by minimum read length (`-l`), maximum read length (`-m`) or average read quality (`-q`).
 
 ```bash
 nanoq -i test.fq -l 1000 -q 10 -m 10000 > reads.fq 
 ```
 
-Read summaries without output can be obtained by directing to `/dev/null` or using the stats flag (`-s / --stats`):
+Read summaries without output can be obtained by directing to `/dev/null` or using the stats flag (`-s`):
 
 ```bash
 nanoq -i test.fq > /dev/null
 nanoq -i test.fq -s
 ```
 
-Read qualities may be excluded from filters and statistics to speed up read iteration in some cases (`-f / --fast`).
+Read qualities may be excluded from filters and statistics to speed up read iteration in some cases (`-f`).
 
 ```bash
 nanoq -i test1.fq.gz -f -s
@@ -141,12 +135,15 @@ A basic read summary is output to `stderr`:
 
 * number of reads
 * number of base pairs
-* read length N50
-* longest and shorted reads
-* mean and median read length
-* mean and median read quality 
+* N50 read length 
+* longest read
+* shorted reads
+* mean read length
+* median read length
+* mean read quality 
+* median read quality
 
[…]
