### 2021-09-27 6afb45a
fix github action syntax
