### 2021-09-27 647d33a
fix github action syntax
