+++ .github/workflows/rust-ci.yaml
@@ -1,5 +1,3 @@
-Based on https://github.com/actions-rs/meta/blob/master/recipes/msrv.md
-
 name: Rust CI
 
 on: [push, pull_request]
