### 2021-09-27 b1e7696
fmt; clippy mods [#18]
