+++ src/needlecast.rs
@@ -1,15 +1,11 @@
-
-
-
-use needletail::{FastxReader, parse_fastx_file, parse_fastx_stdin};
-use std::io::{BufWriter, Write};
-use std::io::{stdout, sink};
 use needletail::errors::ParseError;
+use needletail::{parse_fastx_file, parse_fastx_stdin, FastxReader};
 use std::fs::File;
+use std::io::{sink, stdout};
+use std::io::{BufWriter, Write};
 
 use crate::cli::Cli;
 use crate::utils::CompressionExt;
-
 
 // Niffler output compression adopted from Michael B. Hall - Rasusa (https://github.com/mbhall88/rasusa)
 
@@ -39,7 +35,6 @@ impl NeedleCast {
     /// ```
     #[cfg(not(tarpaulin_include))]
     pub fn new(cli: &Cli) -> Self {
-
         let reader = match &cli.input {
             Some(file) => parse_fastx_file(file).expect("failed to get file reader"),
             None => parse_fastx_stdin().expect("failed to get stdin reader"),
@@ -51,14 +46,14 @@ impl NeedleCast {
                 } else {
                     match cli.output_type {
                         None => Box::new(stdout()),
-                        Some(fmt) => niffler::basic::get_writer(
-                            Box::new(stdout()), fmt, cli.compress_level
-                        ).expect("failed to get compressed stdout writer"),
+                        Some(fmt) => {
+                            niffler::basic::get_writer(Box::new(stdout()), fmt, cli.compress_level)
+                                .expect("failed to get compressed stdout writer")
+                        }
                     }
                 }
-            },
+            }
             Some(output) => {
-
                 let file = File::create(&output).expect("failed to create output file");
                 let file_handle = Box::new(BufWriter::new(file));
 
@@ -66,14 +61,11 @@ impl NeedleCast {
                     None => niffler::Format::from_path(&output),
                     Some(f) => f,
                 };
-                niffler::get_writer(
-                    file_handle, fmt, cli.compress_level
-                ).expect("failed to get compressed file writer")
-            },
+                niffler::get_writer(file_handle, fmt, cli.compress_level)
+                    .expect("failed to get compressed file writer")
+            }
         };
-        NeedleCast {
-            reader, writer
-        }
+        NeedleCast { reader, writer }
     }
     /// Filter reads and store lengths and qualities
     ///
@@ -81,10 +73,10 @@ impl NeedleCast {
     /// and compute average read quality (fastq) records.
     /// Read lengths and qualities are stored in vectors
     /// and returned if no errors are raised
-    /// 
+    ///
     /// # Errors
-    /// 
-    /// If the sequence record cannot be parsed a 
+    ///
+    /// If the sequence record cannot be parsed a
     /// `needletail::errors::ParseError` is returned
     ///
     /// # Example
@@ -99,29 +91,40 @@ impl NeedleCast {
     /// assert_eq!(read_lengths, vec![4]);
     /// assert_eq!(read_quals, vec![40.0]);
     /// ```
-    pub fn filter(&mut self, min_length: u32, max_length: u32, min_quality: f32) -> Result<(Vec<u32>, Vec<f32>), ParseError> {
+    pub fn filter(
+        &mut self,
+        min_length: u32,
+        max_length: u32,
+        min_quality: f32,
+    ) -> Result<(Vec<u32>, Vec<f32>), ParseError> {
         let mut read_lengths: Vec<u32> = vec![];
[…]
