+++ src/utils.rs
@@ -1,17 +1,16 @@
 use anyhow::Result;
 use indoc::eprintdoc;
-use thiserror::Error;
+use std::cmp::Ordering;
 use std::ffi::OsStr;
 use std::path::Path;
+use thiserror::Error;
 
 /// A collection of custom errors relating to the utility components for this package.
 #[derive(Error, Debug, PartialEq)]
 pub enum UtilityError {
-
     /// Indicates an invalid verbosity for summary output
     #[error("{0} is not a valid level of verbosity")]
     InvalidVerbosity(String),
-
 }
 
 // Adopted from Michael B. Hall - Rasusa (https://github.com/mbhall88/rasusa)
@@ -20,7 +19,7 @@ pub trait CompressionExt {
     fn from_path<S: AsRef<OsStr> + ?Sized>(p: &S) -> Self;
 }
 
-/// Attempts to infer the compression type from the file extension. 
+/// Attempts to infer the compression type from the file extension.
 /// If the extension is not known, then Uncompressed is returned.
 impl CompressionExt for niffler::compression::Format {
     fn from_path<S: AsRef<OsStr> + ?Sized>(p: &S) -> Self {
@@ -34,7 +33,6 @@ impl CompressionExt for niffler::compression::Format {
     }
 }
 
-
 /// ReadSet object
 ///
 /// Read set objects are mutable to allow
@@ -42,7 +40,7 @@ impl CompressionExt for niffler::compression::Format {
 ///
 /// * `read_lengths` - a vector of read lengths
 /// * `read_qualities` - a vector of read qualities
-/// 
+///
 #[derive(Debug)]
 pub struct ReadSet {
     read_lengths: Vec<u32>,
@@ -52,7 +50,7 @@ pub struct ReadSet {
 impl ReadSet {
     /// Create a new ReadSet instance
     ///
-    /// Given the verctors of read lengths and 
+    /// Given the verctors of read lengths and
     /// qualities return a mutable ReadSet
     ///
     /// # Example
@@ -76,11 +74,11 @@ impl ReadSet {
     ///     * 2: add length and quality thresholds
     ///     * 3: add top ranked read statistics
     ///
-    /// * `top` - number of top ranking read lengths 
+    /// * `top` - number of top ranking read lengths
     ///     and qualities to show in output
     ///
     /// # Example
-    /// 
+    ///
     /// ```rust
     /// let mut read_set = ReadSet::new(
     ///     !vec[10, 100, 1000], !vec[10.0, 11.0, 12.0]
@@ -88,7 +86,6 @@ impl ReadSet {
     /// read_set.summary(0, 3);
     /// ```
     pub fn summary(&mut self, verbosity: &u64, top: usize) -> Result<(), UtilityError> {
-
         let length_range = self.range_length();
 
         match verbosity {
@@ -106,7 +103,7 @@ impl ReadSet {
                     medianq = self.median_quality(),
                 }
                 Ok(())
-            },
+            }
             &1 | &2 | &3 => {
                 eprintdoc! {"\n
                     Nanoq Read Summary
@@ -132,23 +129,22 @@ impl ReadSet {
                     meanq = self.mean_quality(),
                     medianq = self.median_quality(),
                 }
-                if verbosity > &1  {
+                if verbosity > &1 {
                     self.print_thresholds();
-                } 
+                }
                 if verbosity > &2 {
                     self.print_ranking(top);
                 }
 
                 Ok(())
             }
-            _ => Err(UtilityError::InvalidVerbosity(verbosity.to_string())), 
+            _ => Err(UtilityError::InvalidVerbosity(verbosity.to_string())),
         }
-
     }
[…]
