### 2021-09-27 74212fe
float cmp; clippy satisfied [#18]
