+++ Cargo.toml
@@ -27,6 +27,7 @@ indoc = "1.0"
 needletail = "0.4.1"
 thiserror = "1.0"
 niffler = "2.3"
+float_eq = "0.6.1"
 
 [dev-dependencies]
 assert_cmd = "2.0.1"
