### 2021-09-27 553dbf6
fig title fix [#17]
