### 2021-09-27 esteinig/nanoq#17 @esteinig
In regards to your points @natir

- [x] removed `rust-bio` parser in favor of  `needletail` parser
- [x] added memory usage to benchmarks; added more detailed information for benchmark replication
- [x] separated tables into task and test file benchmarks
- [x] only using `needletail` parser now
- [x] added `rbt sequence-stats` in benchmarks and explanation
- [x] compared output between tools to ensure valid computations (+ added better tests)

`Nanoq` is really fast now  (`--fast`) especially when ignoring quality scores :tada:

Output is compliant between all tools, except that `nanoq` returns an integer for average read length (which is the correct way to return an average of read length in my opinion, as there are no fractional nucleotide bases in reality)
