### 2021-09-27 esteinig/nanoq#18 @esteinig
In regards to your points:
- [x] added continuous integration actions for push, pull requests, and release; added test coverage
- [x] `clippy` compliant code - it was a bit of a mess before, but hopefully much better now
- [x] code split into submodules and documentation added throughout
