+++ README.md
@@ -2,14 +2,14 @@
 
 [![build](https://github.com/esteinig/nanoq/actions/workflows/rust-ci.yaml/badge.svg?branch=master)](https://github.com/esteinig/nanoq/actions/workflows/rust-ci.yaml)
 [![codecov](https://codecov.io/gh/esteinig/nanoq/branch/master/graph/badge.svg?token=1X04YD8YOE)](https://codecov.io/gh/esteinig/nanoq)
-![](https://img.shields.io/badge/version-0.7.0-black.svg)
+![](https://img.shields.io/badge/version-0.8.0-black.svg)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3707754.svg)](https://doi.org/10.5281/zenodo.3707754)
 
 Ultra-fast quality control and summary reports for nanopore reads
 
 ## Overview
 
-**`v0.7.0`**
+**`v0.8.0`**
 
 - [Purpose](#purpose)
 - [Install](#install)
@@ -37,7 +37,7 @@ cargo install nanoq
 #### `Conda`
 
 ```
-conda install -c bioconda nanoq=0.7.0
+conda install -c bioconda nanoq=0.8.0
 ```
 
 #### `Docker`
@@ -101,7 +101,7 @@ done
 ### Parameters
 
 ```
-nanoq 0.7.0
+nanoq 0.8.0
 
 Read filters and summary reports for nanopore data
 
@@ -204,7 +204,7 @@ Tools:
 * `nanostat=1.5.0` 
 * `nanofilt=2.8.0`
 * `filtlong=0.2.1`
-* `nanoq=0.7.0`
+* `nanoq=0.8.0`
 
 Commands used for `stats` task:
 
