+++ meta.yaml
@@ -1,4 +1,4 @@
-{% set version = "0.7.0" %}
+{% set version = "0.8.0" %}
 
 # conda build -c esteinig -c conda-forge .
 
