### 2021-09-30 #2991 @luizirber
Removing the `paused` label, let's get this going again? =]
