### 2021-10-01 8d2cf74
add bin to docs [#17]; update conda recipe [#19]
