+++ meta.yaml
@@ -1,31 +1,35 @@
 {% set version = "0.8.0" %}
-
-# conda build -c esteinig -c conda-forge .
+{% set name = "nanoq" %}
 
 package:
-  name: nanoq
+  name: {{ name }}
   version: {{ version }}
 
 source:
-  path: .
+  url: https://github.com/esteinig/{{ name }}/archive/{{ version }}.tar.gz
+  sha256: a6548374db1439a6b325c872b5fe94d5fdaffbd10a3c6ede6a335f4c3293fa32
 
 build:
   number: 0
 
 requirements:
   build:
-      - {{ compiler('cxx') }}
-  host:
-      - rust>=1.53
-  run:
-      - rust>=1.53
+    - rust >=1.53
+    - {{ compiler('c') }}              # [not win]
+
 test:
   commands:
     - nanoq -h
 
 about:
-  home: 'https://github.com/esteinig/nanoq'
+  home: https://github.com/esteinig/nanoq
   license: MIT
+  summary: Ultra-fast quality control and summary reports for nanopore reads
   license_file: LICENSE
-  summary: 'Minimal but speedy quality control and summaries of nanopore reads'
+
+extra:
+  recipe-maintainers:
+    - esteinig
+    
+
 
