+++ meta.yaml
@@ -1,35 +1,0 @@
-{% set version = "0.8.1" %}
-{% set name = "nanoq" %}
-
-package:
-  name: {{ name }}
-  version: {{ version }}
-
-source:
-  url: https://github.com/esteinig/{{ name }}/archive/{{ version }}.tar.gz
-  sha256: 5debb870f40377b386947db0e43cd4009cb89e04344db86472bbb43b353afe41
-
-build:
-  number: 0
-
-requirements:
-  build:
-    - rust >=1.53
-    - {{ compiler('c') }}              # [not win]
-
-test:
-  commands:
-    - nanoq -h
-
-about:
-  home: https://github.com/esteinig/nanoq
-  license: MIT
-  summary: Ultra-fast quality control and summary reports for nanopore reads
-  license_file: LICENSE
-
-extra:
-  recipe-maintainers:
-    - esteinig
-    
-
-
