### 2021-10-02 esteinig/nanoq#17 @natir
I close this issue last point didn't need to be solve, it's out of scoop now.
