### 2021-10-03 esteinig/nanoq#17 @esteinig
Did you want me to check anything else regarding your last point @natir?
