### 2021-10-04 esteinig/nanoq#17 @natir
First thank you for many works you do, nanoq and paper seems many ways better than before.

Yes and I have some other comments, on paper.

In paper you show, nanoq is faster than any other to get stats and memory usage is reasonable and it's nice. But you didn't compare effect of your filter compare to other, just a plot or table to show read length and read quality after filter. A very interesting bonus, could be if we can see filtering improve down stream analysis, like assembly, species detection, variant calling (make your choice).

Another point, when I read section Applications I feel read a user manual, I think a scientific paper isn't the good place for a user manual. nanoq is able work on active nanopore run is very nice, but you can move this information in another section.

You produce a very expensive comparison of nanoq to other tools, and it's nice but a small text to resume this figure and table could be nice. Just few sentence to say nanoq is faster than any other tools and memory usage is low and constant. I also think it's not need to have all condition in main text, you can just let major result in main text and put other in appendix.

Another recommend change could be replace, long and not so easy to read table, by a scatter plot with computation and memory usage as axes. You lost speedup multiplication but you can add it in caption or in main text and you can put some conditions in the same figure.

I think with all this change paper could be very very nice, but if you @bovee or @luizirber think it's too much change I didn't want block publication.

Just to resume and follow decision on my point (If you wont fix a point just check it)
- [x] Compare filter result
- [x] Remove or move Applications section
- [x] Add comment on result
- [x] Move some result in appendix
- [x] Replace table by scatter plot
