### 2021-10-04 esteinig/nanoq#17 @natir
I also recently discover `seqtk fqchk` have very similar functionally to `nanoq stat`. seqtk is a very famous tools compare nanoq to this one could be nice but just cite it is ok too.
