### 2021-10-04 esteinig/nanoq#17 @esteinig
Appreciate the additional comments on this @natir! I would argue that showing filter results is a bit redundant (it's not very exciting to show the same outputs) and downstream improvement of other applications is not really necessary (and a bit too much for this paper).

Overall these are very basic operations done routinely on nanopore data (from my experience). `Nanoq` is mostly trying to do this faster than other tools. I think a comparison with `seqtk fqchk` would be great in that context.

Agree with the other points, not sure if scatter plot is the right way to show the results to be honest (the data types don't strike me as appropriate for scatter plots, as they generally show relationships between variables). Perhaps an annotated barplot would be better, and only the full dataset benchmark?

I will make tables and figures more legible, condense the benchmarks, and remove the subset benchmarks to the `README` and generally brush this more into paper format.
