### 2021-10-04 esteinig/nanoq#17 @natir
For the filter effect it's ok for me.

About usage of scatter plot to show cpu and memory usage, I agree it's not a perfect fits, with the purpose of plot. But it's really interesting, because you can have different condition on same plot (with different marker), and it's easy to spot the best tools because it's in the left bottom corner. But it's your call.
