### 2021-10-05 4a881d4
fix stats bug [#24]; bump minor version
