+++ src/utils.rs
@@ -95,8 +95,8 @@ impl ReadSet {
                     reads = self.reads(),
                     bases = self.bases(),
                     n50 = self.n50(),
-                    longest = length_range[0],
-                    shortest = length_range[1],
+                    longest = length_range[1],
+                    shortest = length_range[0],
                     mean = self.mean_length(),
                     median = self.median_length(),
                     meanq = self.mean_quality(),
@@ -122,8 +122,8 @@ impl ReadSet {
                     reads = self.reads(),
                     bases = self.bases(),
                     n50 = self.n50(),
-                    longest = length_range[0],
-                    shortest = length_range[1],
+                    longest = length_range[1],
+                    shortest = length_range[0],
                     mean = self.mean_length(),
                     median = self.median_length(),
                     meanq = self.mean_quality(),
