+++ README.md
@@ -2,14 +2,14 @@
 
 [![build](https://github.com/esteinig/nanoq/actions/workflows/rust-ci.yaml/badge.svg?branch=master)](https://github.com/esteinig/nanoq/actions/workflows/rust-ci.yaml)
 [![codecov](https://codecov.io/gh/esteinig/nanoq/branch/master/graph/badge.svg?token=1X04YD8YOE)](https://codecov.io/gh/esteinig/nanoq)
-![](https://img.shields.io/badge/version-0.8.1-black.svg)
+![](https://img.shields.io/badge/version-0.8.2-black.svg)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3707754.svg)](https://doi.org/10.5281/zenodo.3707754)
 
 Ultra-fast quality control and summary reports for nanopore reads
 
 ## Overview
 
-**`v0.8.1`**
+**`v0.8.2`**
 
 - [Purpose](#purpose)
 - [Install](#install)
@@ -38,7 +38,7 @@ nanoq -h
 #### `Conda`
 
 ```
-conda install -c bioconda nanoq=0.8.1
+conda install -c bioconda nanoq=0.8.2
 nanoq -h
 ```
 
@@ -54,7 +54,7 @@ docker run esteinig/nanoq
 Precompiled binaries for Linux and MacOS are attached to the latest release.
 
 ```
-VERSION=0.8.1
+VERSION=0.8.2
 RELEASE=nanoq-${VERSION}-x86_64-unknown-linux-musl.tar.gz
 
 wget https://github.com/esteinig/nanoq/releases/download/${VERSION}/${RELEASE}
@@ -118,7 +118,7 @@ done
 ### Parameters
 
 ```
-nanoq 0.8.1
+nanoq 0.8.2
 
 Read filters and summary reports for nanopore data
 
@@ -221,7 +221,7 @@ Tools:
 * `nanostat=1.5.0` 
 * `nanofilt=2.8.0`
 * `filtlong=0.2.1`
-* `nanoq=0.8.1`
+* `nanoq=0.8.2`
 
 Commands used for `stats` task:
 
