### 2021-10-18 1651769
add seqtk benchmark; rewrite paper [#17]
