+++ paper/plot_benchmarks.py
@@ -51,7 +51,9 @@ def read_data() -> pandas.DataFrame:
             tool = "nanoq-fast"
         elif tool == "rbt":
             tool = "rust-bio-tools"
-        
+        elif tool == "seqtk":
+            tool = "seqtk-fqchk"
+    
 
         df1 = df.iloc[1:11, :].copy()  # cold start
         df1['tool'] = [tool for _ in df1.iterrows()]
@@ -67,7 +69,7 @@ def read_data() -> pandas.DataFrame:
             df2['file'] = ['zymo_full' for _ in df2.iterrows()]
 
             data.append(df2)
-            
+        
 
     return pandas.concat(data)
 
@@ -88,8 +90,8 @@ def plot_data(data: pandas.DataFrame) -> None:
 
     seaborn.barplot(
         y='time', x='file', hue='tool', data=stats_data, order=['zymo_fq', 'zymo_gz'],
-        ax=axes1[0][0], palette=["#de8f05", "#ece133", "#0173b2", "#56b4e9", "#cc78bc"], 
-        hue_order=["nanostat", "nanostat-t8", "nanoq", "nanoq-fast", "rust-bio-tools"]
+        ax=axes1[0][0], palette=["#de8f05", "#ece133", "#029e73", "#0173b2", "#56b4e9", "#cc78bc"], 
+        hue_order=["nanostat", "nanostat-t8", "seqtk-fqchk", "nanoq", "nanoq-fast", "rust-bio-tools"]
     )
     
     seaborn.barplot(
@@ -107,8 +109,8 @@ def plot_data(data: pandas.DataFrame) -> None:
 
     seaborn.barplot(
         y='mem', x='file', hue='tool', data=stats_data, order=['zymo_fq', 'zymo_gz'],
-        ax=axes1[0][1], palette=["#de8f05", "#ece133", "#0173b2", "#56b4e9", "#cc78bc"], 
-        hue_order=["nanostat", "nanostat-t8", "nanoq", "nanoq-fast", "rust-bio-tools"]
+        ax=axes1[0][1], palette=["#de8f05", "#ece133", "#029e73", "#0173b2", "#56b4e9", "#cc78bc"], 
+        hue_order=["nanostat", "nanostat-t8", "seqtk-fqchk", "nanoq", "nanoq-fast", "rust-bio-tools"]
     )
     
     seaborn.barplot(
@@ -144,13 +146,13 @@ def plot_data(data: pandas.DataFrame) -> None:
 
 
     seaborn.barplot(
-        y='time', x='tool', data=stats_data, order=["nanostat", "nanostat-t8", "nanoq", "nanoq-fast", "rust-bio-tools"],
-        ax=axes2[0][0], palette=["#de8f05", "#ece133", "#0173b2", "#56b4e9", "#cc78bc"],
+        y='time', x='tool', data=stats_data, order=["nanostat", "nanostat-t8", "seqtk-fqchk", "nanoq", "nanoq-fast", "rust-bio-tools"],
+        ax=axes2[0][0], palette=["#de8f05", "#ece133",  "#029e73", "#0173b2", "#56b4e9", "#cc78bc"],
     )
     
     seaborn.barplot(
         y='time', x='tool', data=filter_data, order=["nanofilt", "filtlong", "nanoq", "nanoq-fast"],
-        ax=axes2[1][0], palette=["#de8f05", "#029e73", "#0173b2", "#56b4e9"],
+        ax=axes2[1][0], palette=["#de8f05", "#029e73",  "#0173b2", "#56b4e9"],
     )
 
     axes2[0][0].set_xlabel("")
[…]
