### 2021-10-18 esteinig/nanoq#17 @esteinig
Thanks for the patience @natir - I made the following changes to the paper:

* added `seqtk fq`
* commented on benchmark results
* rewrote paper summary
* removed application section
* revised figures but kept barplots
