### 2021-10-18 esteinig/nanoq#17 @natir
Ok I didn't write all of these in this way but it's your paper, information I think are important is present in paper.

Just a point, author of niffler are Luiz Iber and I.
