### 2021-10-19 0159103
revert alt account; added aff for natir, issue [#17]
