+++ paper/paper.md
@@ -19,136 +19,21 @@ authors:
 affiliations:
   - name: The Peter Doherty Institute for Infection and Immunity, The University of Melbourne, Australia
     index: 1
-date: 23 September 2021
+date: 18 October 2021
 bibliography: paper.bib
 ---
 
 # Summary
 
-Nanopore sequencing is now routinely used in a variety of genomics applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. One of the first steps in many workflows is to assess the quality of reads and to obtain summary statistics, as well as to filter fragmented or low quality reads. With increasing throughput on scalable nanopore platforms like `GridION` or `PromethION`, fast quality control of sequence reads and the ability to generate summary statistics on-the-fly are required. When computing minimum summary statistics, `Nanoq` is ~2x faster than `rust-bio-tools` (~5 Gbp and >1 million reads per second) and up to ~440x faster than commonly used alternatives. It offers nanopore-specific quality scores, read filtering options and output compression. `Nanoq` can be applied to data from the public domain, as part of automated pipelines, in streaming applications, or to rapidly check progress of active sequencing runs.
+Nanopore sequencing is now routinely used in a variety of genomics applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. One of the first steps in many workflows is to assess the quality of reads and to obtain summary statistics, as well as to filter fragmented or low quality reads. With increasing throughput on scalable nanopore platforms like `GridION` or `PromethION`, fast quality control of sequence reads and the ability to generate summary statistics on-the-fly are required. Benchmarks indicate that `nanoq` is about as fast as `seqtk fqchk` for small uncompressed data and ~1.5x as fast as `seqtk fqchk` for large uncompressed datasets. Without computing quality scores, `nanoq` is around ~2-3x faster than `rust-bio-tools` for summary statistics and up to ~450x faster than `NanoStats` (> 1.2 million reads per second). In read filtering applications, `nanoq` is faster than other commonly tools (`NanoFilt`, `Filtlong`). Memory consumption is consistent and tends to be lower than alternatives (~3-5x). `Nanoq` offers nanopore-specific quality scores, read filtering options and output compression. it can be applied to data from the public domain, as part of automated pipelines, in streaming applications, or to rapidly check progress of active sequencing runs.
 
 # Statement of need
 
[…]
