### 2021-10-20 b684470
Cleaned up reversion insertion omfg
