### 2021-10-20 esteinig/nanoq#17 @esteinig
Thanks excellent, fixed now. Will close comment but feel free to open or comment if there was anything else
