### 2021-10-24 272df6d
gz fix seqtk; add seqkit benchmark; add data to tables; remove dup fig captions; openjournals/joss-reviews#2991
