+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 # Summary
 
-Nanopore sequencing is now routinely used in a variety of genomics applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. One of the first steps in many workflows is to assess the quality of reads and to obtain summary statistics, as well as to filter fragmented or low quality reads. With increasing throughput on scalable nanopore platforms like `GridION` or `PromethION`, fast quality control of sequence reads and the ability to generate summary statistics on-the-fly are required. Benchmarks indicate that `nanoq` is about as fast as `seqtk fqchk` for small uncompressed data and ~1.5x as fast as `seqtk fqchk` for large uncompressed datasets. Without computing quality scores, `nanoq` is around ~2-3x faster than `rust-bio-tools` for summary statistics and up to ~450x faster than `NanoStats` (> 1.2 million reads per second). In read filtering applications, `nanoq` is faster than other commonly tools (`NanoFilt`, `Filtlong`). Memory consumption is consistent and tends to be lower than alternatives (~3-5x). `Nanoq` offers nanopore-specific quality scores, read filtering options and output compression. it can be applied to data from the public domain, as part of automated pipelines, in streaming applications, or to rapidly check progress of active sequencing runs.
+Nanopore sequencing is now routinely used in a variety of genomics applications, including whole genome assembly [@human_genome] and real-time infectious disease surveillance [@covid]. One of the first steps in many workflows is to assess the quality of reads and to obtain summary statistics, as well as to filter fragmented or low quality reads. With increasing throughput on scalable nanopore platforms like `GridION` or `PromethION`, fast quality control of sequence reads and the ability to generate summary statistics on-the-fly are required. Benchmarks indicate that `nanoq` is as fast as `seqtk` for small datasets (100,000 reads) and ~1.5x as fast for large datasets (3.5 million reads). Without computing quality scores, `nanoq` is around ~2-3x faster than `rust-bio-tools` and `seq-kit stats` for summary statistics, 44x faster than `seqtk` and up to ~450x faster than `NanoStats` (> 1.2 million reads per second). In read filtering applications, `nanoq` is faster than other commonly used tools (`NanoFilt`, `Filtlong`). Memory consumption is consistent and tends to be lower than alternative applications (~5-10x). `Nanoq` offers nanopore-specific quality scores, read filtering options and output compression. It can be applied to data from the public domain, as part of automated pipelines, in streaming applications, or to rapidly check progress of active sequencing runs.
 
 # Statement of need
 
@@ -33,7 +33,7 @@ Nanopore sequencing is now routinely used in a variety of genomics applications,
 
 # Benchmarks
 
[…]
