### 2021-10-24 #2991 @esteinig
* fixed a bug affecting the `seqtk` benchmark (done on uncompressed instead of compressed file)
* added a `seqkit stats` benchmark as it has a 'fast' mode
* added  benchmark data to tables
* removed duplicate figure captions
* updated plots and benchmark text
