### 2021-10-24 f4c52c8
update readme; update citations; openjournals/joss-reviews#2991
