+++ paper/paper.bib
@@ -21,6 +21,21 @@
   doi      =  {10.1093/bioinformatics/btv573}
 }
 
+@ARTICLE{seqkit,
+    doi = {10.1371/journal.pone.0163962},
+    author = {Shen, Wei AND Le, Shuai AND Li, Yan AND Hu, Fuquan},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    title = {SeqKit: A Cross-Platform and Ultrafast Toolkit for FASTA/Q File Manipulation},
+    year = {2016},
+    month = {10},
+    volume = {11},
+    url = {https://doi.org/10.1371/journal.pone.0163962},
+    pages = {1-10},
+    abstract = {FASTA and FASTQ are basic and ubiquitous formats for storing nucleotide and protein sequences. Common manipulations of FASTA/Q file include converting, searching, filtering, deduplication, splitting, shuffling, and sampling. Existing tools only implement some of these manipulations, and not particularly efficiently, and some are only available for certain operating systems. Furthermore, the complicated installation process of required packages and running environments can render these programs less user friendly. This paper describes a cross-platform ultrafast comprehensive toolkit for FASTA/Q processing. SeqKit provides executable binary files for all major operating systems, including Windows, Linux, and Mac OSX, and can be directly used without any dependencies or pre-configurations. SeqKit demonstrates competitive performance in execution time and memory usage compared to similar tools. The efficiency and usability of SeqKit enable researchers to rapidly accomplish common FASTA/Q file manipulations. SeqKit is open source and available on Github at https://github.com/shenwei356/seqkit.},
+    number = {10},
+}
+
 @ARTICLE{covid,
   title     = "Rapid implementation of {SARS-CoV-2} sequencing to investigate
                cases of health-care associated {COVID-19}: a prospective
