+++ README.md
@@ -28,7 +28,7 @@ Ultra-fast quality control and summary reports for nanopore reads
 
 ## Performance
 
-`Nanoq` is about as fast as `seqtk-fqchk` for summary statistics of small uncompressed datasets (+ computes nanopore quality scores) and slightly faster on large uncompressed datasets (~1.5x). In `fast` mode (no quality scores), `nanoq` is faster than `rust-bio-tools` for summary statistics (~2-3x) and faster than commonly used read filtering methods (up to ~450x). Memory consumption is consistent and tends to be lower than alternative tools (~2-5x). 
+`Nanoq` is as fast as `seqtk-fqchk` for summary statistics of small datasets (+ computes nanopore quality scores) and slightly faster on large datasets (~1.5x). In `fast` mode (no quality scores), `nanoq` is faster than `rust-bio-tools` and `seqkit stats` for summary statistics (~2-3x) and faster than other commonly used read filtering methods (up to ~450x). Memory consumption is consistent and tends to be lower than alternative tools (~5-10x). 
 
 ## Install
 
@@ -212,7 +212,7 @@ Top ranking read lengths (bp)
 ## Benchmarks
 
 
-Benchmarks evaluate processing speed and memory consumption of a basic read length filter and summary statistics on the even [Zymo mock community](https://github.com/LomanLab/mockcommunity) (`GridION`) with comparisons to  [`rust-bio-tools`](https://github.com/rust-bio/rust-bio-tools), [`NanoFilt`](https://github.com/wdecoster/nanofilt), [`NanoStat`](https://github.com/wdecoster/nanostat) and [`Filtlong`](https://github.com/rrwick/Filtlong). Time to completion and maximum memory consumption were measured using `/usr/bin/time -f "%e %M"`, speedup is relative to the slowest command in the set. We note that summary statistics from `rust-bio-tools` do not compute read quality score and are therefore comparable to `nanoq-fast`.
+Benchmarks evaluate processing speed and memory consumption of a basic read length filter and summary statistics on the even [Zymo mock community](https://github.com/LomanLab/mockcommunity) (`GridION`) with comparisons to [`rust-bio-tools`](https://github.com/rust-bio/rust-bio-tools), [`seqtk fqchk`](https://github.com/lh3/seqtk), [`seqkit stats`](https://github.com/shenwei356/seqkit), [`NanoFilt`](https://github.com/wdecoster/nanofilt), [`NanoStat`](https://github.com/wdecoster/nanostat) and [`Filtlong`](https://github.com/rrwick/Filtlong). Time to completion and maximum memory consumption were measured using `/usr/bin/time -f "%e %M"`, speedup is relative to the slowest command in the set. We note that summary statistics from `rust-bio-tools` do not compute read quality score and are therefore comparable to `nanoq-fast`.
 
 Tasks:
 
@@ -226,15 +226,16 @@ Tools:
 * `nanofilt 2.8.0`
 * `filtlong 0.2.1`
 * `seqtk 1.3-r126`
+* `seqkit 2.0.0`
 * `nanoq 0.8.2`
 
 Commands used for `stats` task:
 
   * `nanostat` (fq + fq.gz)  --> `NanoStat --fastq test.fq --threads 1` 
-  * `nanostat-t8` (fq + fq.gz) --> `NanoStat --fastq test.fq --threads 8` 
   * `rust-bio` (fq) --> `rbt sequence-stats --fastq < test.fq`
[…]
