### 2021-10-24 #2991 @esteinig
* updated citations in paper
