### 2021-10-24 d8e44d8
spelling; openjournals/joss-reviews#2991
