### 2021-11-11 #2991 @luizirber
@esteinig can you tag a new `nanoq` release (potentially after linking zenodo and github, so new tags get Zenodo DOIs) to serve as the archival version and report the version number and archive DOI here? It is all that is needed to finish this review (so close!)
