### 2021-11-26 #2991 @Kevin-Mattheus-Moerman
@esteinig I have reviewed your paper and the archived data and have the following issues that need your attention:

- [ ] Edit the ZENODO archive such that the meta-data matches the paper.
- The title should be: `Nanoq: ultra-fast quality control for nanopore reads`.
- The author lists should match the paper
- You may also add ORCID profiles to the authors on ZENODO

- [ ] Check throughout for potentially missing "Oxford comma's" E.g. perhaps here? `Nanoq offers nanopore-specific quality scores, read filtering options[,] and output compression.`
- [ ] Check `...they may not immediately applicable...`, looks like that should be `...they may not be immediately applicable...`?
- [ ] Check `...and summary statistics on the even...`. It may be because this is not my domain of expertise but this sentence does not make sense to me. It is mainly the "on the even" part that throws me off. Consider rephrasing.
- [ ] If it is easy to recreate figures 1 and 2 I suggest/recommend (not required) to render them with a larger font size and to export them with a higher resolution.
