### 2022-01-03 a5e28ac
add zenodo data, fix remaining paper; push version openjournals/joss-reviews#2991
