### 2022-01-03 #2991 @esteinig
@arfon yes thanks so much for your patience, the end of last year was a bit hectic for us. please see the new version pushed:

* Zenodo data updated
* oxford commas added throughout manuscript
* spelling checked and fixed
